\name{get_trans_function}
\alias{get_trans_function}

\title{
get_trans_function
}
\description{
Generate a data frame object containing methylation sites and associated transcripts.
}
\usage{
get_trans_function(trans, methyl)
}

\arguments{
  \item{trans}{
Mapped record, generated by the mapToTranscripts() function.
}
  \item{methyl}{
Genomic features, which should be a GRanges object.
}
}

\value{
a data frame object.
}

\examples{
## download the txdb object from UCSC
txdb           <- TxDb.Hsapiens.UCSC.hg19.knownGene

## read genomic features:
data("m6A_methyl_1")

trans                  <- mapToTranscripts(m6A_methyl_1, txdb)
trans_info             <- get_trans_function(trans, m6A_methyl_1)

}
