\name{isoformProb}
\alias{isoformProb}

\title{
isoformProb
}

\description{
Calculte the probability of each genomic feature being located on different isoform transcripts.
}

\usage{
isoformProb(remap_results, includeNeighborDNA = FALSE, lambda = 2)
}

\arguments{
  \item{remap_results}{
Standardized coordinates of genomic features. It can be generated by remapCoord function with  genomic features as input. It should be a list object contains three components, which are an alignment matrix, a width matrix, and a matrix of other associated annotations. For more information please refer to the remapCoord function.
}
  \item{includeNeighborDNA}{
Whether neighborhood DNA regions should be included.
}
  \item{lambda}{
The degree of penalization for information from shorter transcripts.
}
}

\details{
The function returns a data frame, a matrix-like structure whose columns from left to right are row number, feature number, seqnames, feature coordinate (genome-based), strand, transcript ID, and the probability of each genomic feature being located on the corresponding isoform transcript.
}

\value{
A data frame.
}

\examples{
## load remap_results_m6A
data(remap_results_m6A)

isoform_probs <- isoformProb(remap_results_m6A, includeNeighborDNA = TRUE)
}
