\name{MetaTX-package}
\alias{MetaTX-package}
\alias{MetaTX}
\docType{package}
\title{
\packageTitle{MetaTX}
}
\description{
\packageDescription{MetaTX}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MetaTX}
\packageIndices{MetaTX}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MetaTX}

Maintainer: \packageMaintainer{MetaTX}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
