\name{metaTXcount}
\alias{metaTXcount}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
metaTXcount
}
\description{
Calculate the number of features that fall into each bin.
}
\usage{
metaTXcount(remap_results)
}

\arguments{
  \item{remap_results}{
Standardized coordinates of genomic features. It can be generated by remapCoord function with  genomic features as input. It should be a list object contains three components, which are an alignment matrix, a width matrix, and a matrix of other associated annotations. For more information please refer to the remapCoord function.
}
}
\value{
A vector. Each element represents the number of features that fall into the corresponding bin.
}

\examples{
## load remap_results_m6A
data(remap_results_m6A)

feature_counts <- metaTXcount(remap_results_m6A)
}
