\name{metaTXplot}
\alias{metaTXplot}

\title{
metaTXplot
}

\description{
Plot the transcriptomic distribution of genomic features.
}

\usage{
metaTXplot(remap_results, num_bin = 10, includeNeighborDNA = FALSE, comparison = FALSE, lambda = 2, adjust = 0.15, title = "Distribution on mRNA", legend = "sample")
}

\arguments{
  \item{remap_results}{
Standardized coordinates of genomic features. It can be generated by remapCoord function with  genomic features as input. It should be a list object contains three components, which are an alignment matrix, a width matrix, and a matrix of other associated annotations. For more information please refer to the remapCoord function.
}
  \item{num_bin}{
The number of bins into which an mRNA component is divided.
}
  \item{includeNeighborDNA}{
Whether neighborhood DNA regions should be included.
}
  \item{comparison}{
Whether results of direct estimation method should be included.
}
  \item{lambda}{
The degree of penalization for information from shorter transcripts.
}
  \item{adjust}{
A multiplicate bandwidth adjustment. This makes it possible to adjust the bandwidth while still using the a bandwidth estimator. For example, adjust = 1/2 means use half of the default bandwidth.
}
  \item{title}{
The title of the plot.
}
  \item{legend}{
The legend of the plot.
}
}
\details{
This function plots the transcriptomic distribution of genomic features. It is designed for a fast usage of the MetaTX package without the needs to go into the details.
}
\value{
A figure showing the transcriptomic distribution of the genomic features will be generated.
}

\examples{
## load remap_results_m6A
data(remap_results_m6A)

## download the txdb object from UCSC
txdb  <- TxDb.Hsapiens.UCSC.hg19.knownGene

## plot
p1 <- metaTXplot(remap_results_m6A, txdb,  
           includeNeighborDNA = TRUE, 
           comparison = TRUE)
p1
}
