\name{remap_results_m6A}
\alias{remap_results_m6A}
\docType{data}
\title{
remap_results_m6A
}
\description{
The remap_results_m6A is a container for storing remapping results of the feature set m6A_methyl_1
}
\usage{data("remap_results_m6A")}

\examples{
## load remap_results_m6A
data(remap_results_m6A)

}
\keyword{datasets}
