\name{remap_results_m6A_1}
\alias{remap_results_m6A_1}
\docType{data}
\title{
remap_results_m6A_1
}
\description{
The remap_results_m6A is a container for storing remapping results of the feature set remap_results_m6A_1.
}
\examples{
## load remap_results_m6A_1
data(remap_results_m6A_1)
}
\keyword{datasets}
